refactor(eurora/prosemirror-view): initial migration of launcher and related components
