+++ apps/docs/src/routes/components/launcher/+page.svelte
@@ -4,7 +4,7 @@
 	import SiGoogledrive from '@icons-pack/svelte-simple-icons/icons/SiGoogledrive';
 
 	// import { LauncherNative as Launcher } from '@eurora/launcher';
-	import * as Launcher from '@eurora/ui/custom-components/launcher/index';
+	import * as Launcher from '@eurora/prosemirror-view/launcher';
 	import { transcriptExtension } from '@eurora/ext-transcript/index';
 	import { Editor as ProsemirrorEditor } from '@eurora/prosemirror-core/index';
 	let editorRef: ProsemirrorEditor | undefined = $state(undefined);
