+++ packages/prosemirror/view/package.json
@@ -10,6 +10,14 @@
 		"**/*.css"
 	],
 	"exports": {
+		".": {
+			"import": "./dist/index.js",
+			"types": "./dist/index.d.ts"
+		},
+		"./launcher": {
+			"import": "./dist/launcher/index.js",
+			"types": "./dist/launcher/index.d.ts"
+		},
 		"./*": {
 			"import": "./dist/*.js",
 			"types": "./dist/*.d.ts"
@@ -42,6 +50,7 @@
 		"@sveltejs/kit": "catalog:svelte",
 		"@sveltejs/package": "^2.5.0",
 		"@eurora/prosemirror-core": "workspace:*",
+		"@eurora/ui": "workspace:*",
 		"vite": "catalog:",
 		"vitest": "catalog:",
 		"prosemirror-commands": "^1.5.2",
@@ -55,6 +64,12 @@
 		"prosemirror-example-setup": "^1.1.2",
 		"@eurora/typescript-config": "workspace:*",
 		"prosemirror-dropcursor": "^1.8.2",
-		"prosemirror-gapcursor": "^1.3.2"
+		"prosemirror-gapcursor": "^1.3.2",
+		"bits-ui": "^2.9.6",
+		"@lucide/svelte": "^0.542.0",
+		"clsx": "^2.1.1",
+		"tailwind-merge": "^3.3.1",
+		"@storybook/addon-svelte-csf": "^5.0.8",
+		"storybook": "^8.6.12"
 	}
 }
