+++ packages/prosemirror/view/src/lib/index.ts
@@ -1,0 +1,2 @@
+// Export launcher components
+export * from './launcher/index.js';
