+++ packages/prosemirror/view/src/lib/launcher/command-item.svelte
@@ -1,15 +1,15 @@
 <script lang="ts">
 	import { Command as CommandPrimitive } from 'bits-ui';
-	import { cn } from '$lib/utils.js';
+	import { cn } from '$lib/utils';
 
 	let {
 		ref = $bindable(null),
 		class: className,
 		...restProps
-	}: CommandPrimitive.LinkItemProps = $props();
+	}: CommandPrimitive.ItemProps = $props();
 </script>
 
-<CommandPrimitive.LinkItem
+<CommandPrimitive.Item
 	class={cn(
 		'aria-selected:bg-accent aria-selected:text-accent-foreground relative flex cursor-default items-center gap-2 rounded-sm px-2 py-1.5 text-sm outline-none select-none data-[disabled]:pointer-events-none data-[disabled]:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0',
 		className,
