+++ packages/prosemirror/view/src/lib/launcher/command-shortcut.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-	import type { WithElementRef } from '$lib/utils.js';
+	import type { WithElementRef } from '$lib/utils';
 	import type { HTMLAttributes } from 'svelte/elements';
-	import { cn } from '$lib/utils.js';
+	import { cn } from '$lib/utils';
 
 	let {
 		ref = $bindable(null),
