+++ packages/prosemirror/view/src/lib/launcher/input-area.svelte
@@ -12,7 +12,7 @@
 </script>
 
 <script lang="ts">
-	import { cn } from '$lib/utils.js';
+	import { cn } from '$lib/utils';
 	import { onMount } from 'svelte';
 
 	let {
