+++ packages/prosemirror/view/src/stories/launcher/Command.stories.svelte
@@ -1,6 +1,6 @@
 <script module lang="ts">
-	import CommandRoot from '$lib/custom-components/launcher/command.svelte';
-	import * as Command from '$lib/custom-components/launcher/index.js';
+	import CommandRoot from '$lib/launcher/command.svelte';
+	import * as Command from '$lib/launcher/index.js';
 	import { defineMeta, type StoryContext, type Args } from '@storybook/addon-svelte-csf';
 
 	const { Story } = defineMeta({
@@ -28,7 +28,7 @@
 </script>
 
 <script lang="ts">
-	import { StorybookContainer } from '$lib/custom-components/storybook-container/index.js';
+	import { StorybookContainer } from '@eurora/ui/storybook-container/index.js';
 	import {
 		Search,
 		Calculator,
