+++ packages/prosemirror/view/src/stories/launcher/CommandGroup.stories.svelte
@@ -1,5 +1,5 @@
 <script module lang="ts">
-	import { CommandGroup } from '$lib/custom-components/launcher/index.js';
+	import { CommandGroup } from '$lib/launcher/index.js';
 	import { defineMeta, type StoryContext, type Args } from '@storybook/addon-svelte-csf';
 
 	const { Story } = defineMeta({
@@ -26,8 +26,8 @@
 </script>
 
 <script lang="ts">
-	import * as Command from '$lib/custom-components/launcher/index.js';
-	import { StorybookContainer } from '$lib/custom-components/storybook-container/index.js';
+	import * as Command from '$lib/launcher/index.js';
+	import { StorybookContainer } from '@eurora/ui/storybook-container/index.js';
 	import {
 		Calendar,
 		Calculator,
