+++ packages/ui/package.json
@@ -28,7 +28,6 @@
 		"@eslint/js": "^9.35.0",
 		"@eurora/shared": "workspace:*",
 		"@eurora/eslint-config": "workspace:*",
-		"@eurora/prosemirror-core": "workspace:*",
 		"@eurora/typescript-config": "workspace:*",
 		"@internationalized/date": "^3.9.0",
 		"@lucide/svelte": "^0.542.0",
